Refactor view wishlist template and add share button
